+++ spec/dummy/app/models/book.rb
@@ -4,10 +4,10 @@ class Book < ActiveRecord::Base
   store_attributes :data do
     author String
     title String, default: "Great Voyage"
-    available Axiom::Types::Boolean, default: true
+    available Boolean, default: true
     copies Integer, default: 0
   end
-  store_attribute :data, :hardcover, Axiom::Types::Boolean
+  store_attribute :data, :hardcover, Boolean
   store_attribute :data, :alt_name
 
 end